Node 24 CI の polyfill 検証を安定化する
